parent/uav/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
git+https://github.com/dronekit/dronekit-python.git@ed132c#egg=dronekit
dronekit-sitl==3.2.0
future==0.17.1
lxml==4.3.0               # via pymavlink
mavproxy==1.6.4
monotonic==1.5
psutil==5.6.6             # via dronekit-sitl
pymavlink==2.3.4
pyserial==3.4             # via mavproxy
simplejson==3.16.0
six==1.12.0               # via dronekit-sitl
